Bumped up version and kept CHANGELOG up to date
The latest version (v0.1.3) of this pack is tagged at the point that #6
was merged. So it doesn't contain the changes from #7 to #9.And those
changes have not been written at the CHANGELOG.
This commit bump up version to v0.1.4 and keep track of descriptions of
those features which were merged after v0.1.3 to the CHANGELOG as the
ones of v0.1.4.